moved data folder within src to populate initial cards
