+++ src/components/App/App.js
@@ -3,6 +3,8 @@ import './App.css';
 import Search from '../Search/Search';
 import CardContainer from '../CardContainer/CardContainer';
 import CompareContainer from '../CompareContainer/CompareContainer';
+import DistrictRepository from '../../helper';
+import kinderData from '../../data/kindergartners_in_full_day_program';
 
 class App extends Component {
   constructor(props) {
@@ -10,6 +12,11 @@ class App extends Component {
     this.state = {
       districts: []
     }
+  }
+
+  componentDidMount() {
+    const districts = new DistrictRepository(kinderData);
+    this.setState({ districts: districts });
   }
 
   render() {
