+++ src/test/unit/iteration-4.test.js
@@ -1,5 +1,5 @@
 import DistrictRepository from '../../helper';
-import kinderData from '../../../data/kindergartners_in_full_day_program';
+import kinderData from '../../data/kindergartners_in_full_day_program';
 
 describe('DistrictRepository iteration 0', () =>  {
   const district = new DistrictRepository(kinderData);
